requirements-rpi.txt
# CONTROLLER
wheel
# asyncio-nats-client
nats-py

RPi.GPIO
pi-rc522
tinydb
tinydb-serialization
pyserial

setuptools_rust
